+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Fixed
-- Crash(?) (not as much a crash, more like a car running out of fuel in the middle of nowhere) with FancyMenu and C2ME *together only*
+- Crash on resource reload with F3+T